+++ src/layouts/window/app.vue
@@ -10,22 +10,39 @@ import { TooltipProvider } from '@/components/ui/tooltip';
 import { Toaster } from '@/components/ui/toaster';
 import { InfusionWrapper } from '@/components/ui/infusion';
 import { ChangelogDialog } from '@/modules/changelog';
+import { useUserSettingsStore } from '@/stores/storage/user-settings';
+
+const userSettingsStore = useUserSettingsStore();
 </script>
 
 <template>
-  <InfusionWrapper />
-  <Toaster />
-  <ChangelogDialog />
-  <TooltipProvider>
-    <NavSidebar />
-    <div class="app-layout__main">
-      <WindowToolbar />
-      <RouterView />
-    </div>
-  </TooltipProvider>
+  <div
+    class="app-layout"
+    :style="{ '--font': userSettingsStore.defaultFontFamily }"
+  >
+    <InfusionWrapper />
+    <Toaster />
+    <ChangelogDialog />
+    <TooltipProvider>
+      <NavSidebar />
+      <div class="app-layout__main">
+        <WindowToolbar />
+        <RouterView />
+      </div>
+    </TooltipProvider>
+  </div>
 </template>
 
 <style>
+.app-layout {
+  --default-font-family: var(--font, system-ui);
+
+  display: flex;
+  width: 100%;
+  height: 100%;
+  font-family: var(--default-font-family);
+}
+
 .app-layout__main {
   position: relative;
   width: 100%;